+++ lib/netutil/conn_pool.go
@@ -236,13 +236,13 @@ func (cp *ConnPool) closeIdleConns() {
 	// Close connections, which were idle for more than 120 seconds.
 	// This should reduce the number of connections after sudden spikes in query rate.
 	// See https://github.com/VictoriaMetrics/VictoriaMetrics/issues/2508
-	deadline := fasttime.UnixTimestamp() - 120
+	deadline := fasttime.UnixTimestamp() - 60
 	var activeConns []connWithTimestamp
 	var closeConns []connWithTimestamp
 	cp.mu.Lock()
 
-	// fast path, if there are less than 3 connections in the pool.
-	if len(cp.conns) < 3 {
+	// fast path, if there are less than 2 connections in the pool.
+	if len(cp.conns) < 2 {
 		cp.mu.Unlock()
 		return
 	}
@@ -256,7 +256,10 @@ func (cp *ConnPool) closeIdleConns() {
 		}
 	}
 	for _, c := range closeConns {
-		if len(activeConns) < 3 {
+		// It can limit the number of broken connections remaining after the idle timeout to 1,
+		// preventing failed requests and partial responses caused by too many broken connections after the idle timeout.
+		// Please see https://github.com/VictoriaMetrics/VictoriaMetrics/issues/10314.
+		if len(activeConns) < 1 {
 			activeConns = append(activeConns, c)
 			continue
 		}